# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PM Gestion API
  version: 1.1.0
  description: API de gestion de projets — Authentification JWT, RBAC 2-tier, real-time Socket.IO
  contact:
    name: TripleV
    url: https://github.com/TripleV-coder/Project-Manager
servers:
  - url: http://localhost:3000/api
    description: Development
  - url: '{baseUrl}/api'
    description: Production
    variables:
      baseUrl:
        default: https://your-domain.com
security:
  - BearerAuth: []
  - CookieAuth: []
...
paths:
...
  /users/{id}:
    get:
      tags: [Users]
      summary: Get user details
      parameters: [{ $ref: '#/components/parameters/ResourceId' }]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
    put:
      tags: [Users]
      summary: Update user
      parameters: [{ $ref: '#/components/parameters/ResourceId' }]
      responses:
        '200': { description: User updated }
    delete:
      tags: [Users]
      summary: Deactivate user
      parameters: [{ $ref: '#/components/parameters/ResourceId' }]
      responses:
        '204': { description: User deactivated }

  # ── Comments ──────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        pattern: '^[a-f0-9]{24}$'
      description: MongoDB ObjectId
...
  schemas:
    User:
      type: object
      properties:
        _id: { type: string }
        nom_complet: { type: string }
        email: { type: string, format: email }
        role: { $ref: '#/components/schemas/Role' }
        statut: { type: string, enum: [Actif, Inactif, Suspendu] }
        département: { type: string }
        poste: { type: string }
        compétences: { type: array, items: { type: string } }
    Role:
      type: object
      properties:
        _id: { type: string }
        nom: { type: string }
        description: { type: string }
        permissions:
          type: object
          description: 32 atomic boolean permissions
...